商品削除機能の実装修正1

+++ app/controllers/items_controller.rb
@@ -36,7 +36,7 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
+    item.destroy if user_signed_in? && current_user == item.user
     redirect_to root_path
   end
